clean up allocators
