+++ sgl-kernel/csrc/common_extension_rocm.cc
